bot/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
py-cord==2.4.0
